requirements.txt
# Core chess functionality
python-chess>=1.9.0
stockfish>=3.28.0

# Natural Language Processing
spacy>=3.5.0
nltk>=3.8.1
transformers>=4.30.0

# Text Generation
tracery>=0.1.1
markovify>=0.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
pymongo>=4.3.0

# Visualization
pygame>=2.5.0
plotly>=5.14.0
networkx>=3.1.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.65.0

# Testing
pytest>=7.3.1
pytest-cov>=4.0.0
